+++ tests/Tests/Unit/GapicClientTraitTest.php
@@ -873,7 +873,7 @@ class GapicClientTraitTest extends TestCase
         }
         $expectedProperties = [
             'serviceName' => 'test.interface.v1.api',
-            'agentHeader' => AgentHeader::buildAgentHeader([]),
+            'agentHeader' => AgentHeader::buildAgentHeader([]) + ['User-Agent' => ['gcloud-php-legacy/']],
             'retrySettings' => $expectedRetrySettings,
         ];
         return [
@@ -1358,7 +1358,8 @@ class GapicClientTraitTest extends TestCase
                 $this->isInstanceOf(Call::class),
                 $this->equalTo([
                     'headers' => AgentHeader::buildAgentHeader([]) + [
-                        'X-Goog-User-Project' => [$quotaProject]
+                        'X-Goog-User-Project' => [$quotaProject],
+                        'User-Agent' => ['gcloud-php-legacy/']
                     ],
                     'credentialsWrapper' => $credentialsWrapper
                 ])
@@ -1390,6 +1391,16 @@ class GapicClientTraitTest extends TestCase
         // verify scopes are not set by default
         $defaultOptions = $client->call('buildClientOptions', [[]]);
         $this->assertArrayNotHasKey('scopes', $defaultOptions['credentialsConfig']);
+
+        // verify scopes are set when a custom api endpoint is used
+        $defaultOptions = $client->call('buildClientOptions', [[
+            'apiEndpoint' => 'www.someotherendpoint.com',
+        ]]);
+        $this->assertArrayHasKey('scopes', $defaultOptions['credentialsConfig']);
+        $this->assertEquals(
+            $client::$serviceScopes,
+            $defaultOptions['credentialsConfig']['scopes']
+        );
 
         // verify user-defined scopes override default scopes
         $defaultOptions = $client->call('buildClientOptions', [[
@@ -1450,36 +1461,6 @@ class GapicClientTraitTest extends TestCase
 
         $client->call('startCall', ['method.name', 'decodeType', [
             'audience' => 'custom-audience',
-        ]]);
-    }
-
-    public function testDefaultAudienceWithCustomApiEndpoint()
-    {
-        $retrySettings = $this->prophesize(RetrySettings::class);
-        $credentialsWrapper = $this->prophesize(CredentialsWrapper::class)
-            ->reveal();
-        $transport = $this->prophesize(TransportInterface::class);
-        $transport
-            ->startUnaryCall(
-                Argument::any(),
-                [
-                    'audience' => 'https://service-address/',
-                    'headers' => [],
-                    'credentialsWrapper' => $credentialsWrapper,
-                ]
-            )
-            ->shouldBeCalledOnce();
-
-        $client = new GapicClientTraitDefaultScopeAndAudienceStub();
-        $client->set('credentialsWrapper', $credentialsWrapper);
-        $client->set('agentHeader', []);
-        $client->set(
-            'retrySettings',
-            ['method.name' => $retrySettings->reveal()]
-        );
-        $client->set('transport', $transport->reveal());
-        $client->call('startCall', ['method.name', 'decodeType', [
-            'apiEndpoint' => 'www.someotherendpoint.com',
         ]]);
     }
 
@@ -1840,6 +1821,25 @@ class GapicClientTraitTest extends TestCase
             new MockRequest(),
         ])->wait();
     }
+
+    public function testSurfaceAgentHeaders()
+    {
+        // V1 does not contain new headers
+        $client = new GapicClientTraitRestOnly([
+            'gapicVersion' => '0.0.2',
+        ]);
+        $agentHeader = $client->getAgentHeader();
+        $this->assertStringContainsString(' gapic/0.0.2 ', $agentHeader['x-goog-api-client'][0]);
+        $this->assertEquals('gcloud-php-legacy/0.0.2', $agentHeader['User-Agent'][0]);
+
+        // V2 contains new headers
+        $client = new GapicV2SurfaceClient([
+            'gapicVersion' => '0.0.1',
+        ]);
+        $agentHeader = $client->getAgentHeader();
+        $this->assertStringContainsString(' gapic/0.0.1 ', $agentHeader['x-goog-api-client'][0]);
+        $this->assertEquals('gcloud-php-new/0.0.1', $agentHeader['User-Agent'][0]);
+    }
 }
 
 class GapicClientTraitStub
@@ -1985,6 +1985,11 @@ class GapicClientTraitRestOnly
     {
         return 'rest';
     }
+
+    public function getAgentHeader()
+    {
+        return $this->agentHeader;
+    }
 }
 
 class GapicClientTraitOperationsStub extends GapicClientTraitStub
@@ -2028,6 +2033,11 @@ abstract class GapicV2SurfaceBaseClient
                 ]
             ],
         ];
+    }
+
+    public function getAgentHeader()
+    {
+        return $this->agentHeader;
     }
 }